+++ frontend/lib/presentation/boilerplate/navigation_bar.dart
@@ -1,0 +1,57 @@
+import 'package:flutter/material.dart';
+
+class MyNavigationBar extends StatefulWidget {
+  const MyNavigationBar({super.key});
+
+  @override
+  _MyNavigationBarState createState() => _MyNavigationBarState();
+}
+
+class _MyNavigationBarState extends State<MyNavigationBar> {
+  int _selectedIndex = 0; // Track selected tab index
+
+  // Handle tab changes
+  void _onItemTapped(int index) {
+    setState(() {
+      _selectedIndex = index;
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return BottomNavigationBar(
+      backgroundColor: Colors.white, // Custom background
+      currentIndex: _selectedIndex, // Active tab
+      onTap: _onItemTapped, // Tab change callback
+
+      // Active/Inactive icon colors
+      selectedItemColor: Theme.of(context).colorScheme.primary,
+      unselectedItemColor: Colors.grey,
+
+      // Font sizes and custom label styles
+      selectedFontSize: 14,
+      unselectedFontSize: 12,
+      selectedLabelStyle: const TextStyle(fontWeight: FontWeight.bold),
+      unselectedLabelStyle: const TextStyle(fontWeight: FontWeight.w400),
+
+      // Navigation Items
+      items: const [
+        BottomNavigationBarItem(
+          icon: Icon(Icons.home_outlined),
+          activeIcon: Icon(Icons.home),
+          label: "Home",
+        ),
+        BottomNavigationBarItem(
+          icon: Icon(Icons.search_outlined),
+          activeIcon: Icon(Icons.search),
+          label: "Search",
+        ),
+        BottomNavigationBarItem(
+          icon: Icon(Icons.person_outline),
+          activeIcon: Icon(Icons.person),
+          label: "Profile",
+        ),
+      ],
+    );
+  }
+}